Fix: Check if attribute is nullable in associated value before deletion.

+++ lib/eMapper/Reflection/Profile/Association/OneToMany.php
@@ -158,10 +158,23 @@ class OneToMany extends Association {
 	
 	public function delete($mapper, $foreignKey) {
 		$manager = $mapper->buildManager($this->profile);
-		$related = $manager->find(Attr::__callstatic($this->attribute->getValue())->eq($foreignKey));
+		$attr = $this->attribute->getValue();
+		$related = $manager->find(Attr::__callstatic($attr)->eq($foreignKey));
 		
-		foreach ($related as $value) {
-			$manager->delete($value);
+		//determine if the attribute is nullable
+		$entityProfile = Profiler::getClassProfile($this->profile);
+		$property = $entityProfile->getProperty($attr);
+		
+		if ($property->isNullable()) {
+			$query = sprintf("UPDATE %s SET %s = NULL WHERE %s = %s", $entityProfile->getReferredTable(), $property->getColumn(), $property->getColumn(), $foreignKey);
+			$mapper->sql($query);
+		}
+		else {
+			$related = $manager->find(Attr::__callstatic($attr)->eq($foreignKey));
+			
+			foreach ($related as $value) {
+				$manager->delete($value);
+			}
 		}
 	}
